refactor: TourController -> NewTourController
