+++ lib/app/config/routes/route_path.dart
@@ -13,7 +13,7 @@ class RoutePath {
   static const String term = "/term";
   static const String setProfile = "/set_profile";
 
-  // Add Tour
+  // Add New Tour
   static const String selectDate = "/select_date";
   static const String selectTitle = "/select_title";
   static const String selectRegion = "/select_region";
