+++ lib/presentation/new_tour/controller/new_tour_binding.dart
@@ -1,12 +1,12 @@
 import 'package:get/get.dart';
 
 import 'calendar_controller.dart';
-import 'tour_controller.dart';
+import 'new_tour_controller.dart';
 
-class TourBinding extends Bindings {
+class NewTourBinding extends Bindings {
   @override
   void dependencies() {
-    Get.lazyPut(() => TourController(), fenix: true);
+    Get.lazyPut(() => NewTourController(), fenix: true);
     Get.lazyPut(() => CalendarController(), fenix: true);
   }
 }
